Fix test inconsistency due to weaker consistency
Milvus supports different levels of consistency depending on the demands
of the business case. The default setting is Bounded staleness, which
provides relatively good syncronization between replicas and much faster
inference times. This is undesirable for unit tests, however, because it
can result in stochastic behavior for some tests versus others.

Add a new customization parameter to the model definitions that allow
for specifying a different type of consistency for the schema.